+++ dds.go
@@ -132,7 +132,7 @@ func (d *DDS) Write(w io.Writer) error {
 
 	if !d.Format.isCompressed() {
 		switch d.Format {
-		case FormatRGBA8, FormatBGRA8, FormatBGRX8, FormatR8, FormatRG8, FormatRGB10A2, FormatR8S, FormatRG8S:
+		case FormatRGBA8, FormatBGRA8, FormatBGRX8, FormatR8, FormatRG8, FormatRGB10A2, FormatR8S, FormatRG8S, FormatA8:
 		default:
 			return ErrUnsupportedFormat
 		}
@@ -294,6 +294,13 @@ func (d *DDS) Write(w io.Writer) error {
 		hdr.PitchOrLinearSize = u32(d.Width * 2)
 		dx10Format = 51 // DXGI_FORMAT_R8G8_SNORM
 
+	case FormatA8:
+		hdr.Flags |= DDSFlagPitch
+		hdr.PixelFormat.Flags = DDSPFFourCC
+		hdr.PixelFormat.FourCC = makeFourCC('D', 'X', '1', '0')
+		hdr.PitchOrLinearSize = u32(d.Width)
+		dx10Format = 65 // DXGI_FORMAT_A8_UNORM
+
 	default:
 		return ErrUnsupportedDDSFormat
 	}
@@ -372,6 +379,8 @@ func ddsFormatFromHeader(r io.Reader, header *DDSHeader) (Format, *DDSHeaderDX10
 			return FormatR8, &dx10, nil
 		case 63: // DXGI_FORMAT_R8_SNORM
 			return FormatR8S, &dx10, nil
+		case 65: // DXGI_FORMAT_A8_UNORM
+			return FormatA8, &dx10, nil
 		case 71, 72: // DXGI_FORMAT_BC1_UNORM, DXGI_FORMAT_BC1_UNORM_SRGB
 			return FormatBC1, &dx10, nil
 		case 74, 75: // DXGI_FORMAT_BC2_UNORM, DXGI_FORMAT_BC2_UNORM_SRGB